compiles, works (displaying just english-names)
GHCJS was being weired with data declarations, so I moved those to a
backend-only types file and now its fine. *shrug?*

Previous functionality is back, but now data is read from the EOLAS
files and can be searched on more richly.  Next steps will be editing
the front-end so that multi-field searches can be performed and
multi-field results can be displayed.  The back-end should be ready to
serve those.