better matching
The typed string can be matched to any word in result strings, not just
the start of the string. This was enabled to make matching over the
'classification' field work, even though we're not actually using that.

I made it work in kind of a hack-y way using regexes. The better
solution would be to make the results lists of strings, with each string
representing a matchable word or phrase? that would be more work though!